Namespace the controllers and routes
everything is now from /api/v1/
