+++ app/controllers/api/v1/episodes_controller.rb
@@ -1,0 +1,18 @@
+# frozen_string_literal: true
+
+module Api
+  module V1
+    # Define the Espiode Controller
+    class EpisodesController < ApplicationController
+      def index
+        episodes = Episode.all
+        render json: episodes
+      end
+
+      def show
+        episode = Episode.find(params[:id])
+        render json: episode
+      end
+    end
+  end
+end
