--- app/controllers/episodes_controller.rb
@@ -1,14 +1,0 @@
-# frozen_string_literal: true
-
-# Define the Espiode Controller
-class EpisodesController < ApplicationController
-  def index
-    episodes = Episode.all
-    render json: episodes
-  end
-
-  def show
-    episode = Episode.find(params[:id])
-    render json: episode
-  end
-end
