--- app/controllers/seasons_controller.rb
@@ -1,14 +1,0 @@
-# frozen_string_literal: true
-
-# Define the Season Controller
-class SeasonsController < ApplicationController
-  def index
-    seasons = Season.all
-    render json: seasons
-  end
-
-  def show
-    season = Season.find(params[:id])
-    render json: season
-  end
-end
